Decouple quarkus-spring-web from RESTEasy Classic
This is prerequisite work for being able to run
Spring Web endpoints on RESTEasy Reactive in addition
to RESTEasy Classic.

The decoupling is achieved by using conditional
dependencies and moving RESTEasy specific code
to a new extension
