+++ extensions/spring-web/core/deployment/src/main/java/io/quarkus/spring/web/deployment/TypesUtil.java
