+++ extensions/spring-web/core/deployment/src/test/java/io/quarkus/spring/web/test/SimpleSpringController.java
