+++ extensions/spring-web/core/runtime/src/main/resources/META-INF/services/io.quarkus.runtime.test.TestHttpEndpointProvider
