+++ extensions/spring-web/resteasy-classic/runtime/src/main/java/io/quarkus/spring/web/runtime/ResteasyClassicResponseContentTypeResolver.java
@@ -1,0 +1,21 @@
+package io.quarkus.spring.web.runtime;
+
+import java.util.List;
+
+import javax.ws.rs.core.Variant;
+
+import org.jboss.resteasy.core.request.ServerDrivenNegotiation;
+
+import io.quarkus.spring.web.runtime.common.AbstractResponseContentTypeResolver;
+
+@SuppressWarnings("unused")
+public class ResteasyClassicResponseContentTypeResolver extends AbstractResponseContentTypeResolver {
+
+    @Override
+    protected Variant negotiateBestMatch(List<String> acceptHeaders, List<Variant> variants) {
+        ServerDrivenNegotiation negotiation = new ServerDrivenNegotiation();
+        negotiation.setAcceptHeaders(acceptHeaders);
+
+        return negotiation.getBestMatch(variants);
+    }
+}
