+++ extensions/spring-web/resteasy-classic/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,5 @@
+---
+artifact: ${project.groupId}:${project.artifactId}:${project.version}
+name: "Spring Web RESTEasy Classic"
+metadata:
+  unlisted: true
